// src/enums/post-form-field.enum.ts
export enum PostFormField {
  Title = 'title',
  Content = 'content',
  Slug = 'slug',
  Status = 'status',
  CategoryIds = 'categoryIds',
  ImageKey = 'imageKey',
  ImageUrl = 'imageUrl',
  ImageMimeType = 'imageMimeType',
  ImageSizeBytes = 'imageSizeBytes',
  ImageOriginalFileName = 'imageOriginalFileName',
  ImageAlt = 'imageAlt',
}